Bugfix: result of median is not truncated anymore

+++ src/math_utility.cpp
@@ -4,12 +4,12 @@
 
 namespace math_utility
 {
-  
-  double calc_sum(const std::vector<double>& toCalculateFrom)
+
+  double calc_sum(const std::vector<double> &toCalculateFrom)
   {
     double total_sum{0.0};
 
-    for (const auto& number : toCalculateFrom)
+    for (const auto &number : toCalculateFrom)
     {
       total_sum += number;
     }
@@ -17,20 +17,20 @@ namespace math_utility
     return total_sum;
   }
 
-  double calc_avg(const std::vector<double>& toCalculateFrom)
+  double calc_avg(const std::vector<double> &toCalculateFrom)
   {
     double average{0.0};
 
-    for (const auto& number : toCalculateFrom)
+    for (const auto &number : toCalculateFrom)
     {
       average += number;
     }
-    
+
     double divider = static_cast<double>(toCalculateFrom.size());
     return average / divider;
   }
 
-  double calc_mult(const std::vector<double>& toCalculateFrom)
+  double calc_mult(const std::vector<double> &toCalculateFrom)
   {
     if (toCalculateFrom.size() == 0)
     {
@@ -39,15 +39,15 @@ namespace math_utility
 
     double product{1.0};
 
-    for (const auto& number : toCalculateFrom)
+    for (const auto &number : toCalculateFrom)
     {
       product *= number;
     }
-    
+
     return product;
   }
 
-  double calc_min(const std::vector<double>& toCalculateFrom)
+  double calc_min(const std::vector<double> &toCalculateFrom)
   {
     if (toCalculateFrom.size() == 0)
     {
@@ -56,15 +56,15 @@ namespace math_utility
 
     double min{toCalculateFrom.at(0)};
 
-    for (const auto& number : toCalculateFrom)
+    for (const auto &number : toCalculateFrom)
     {
       min = std::min(number, min);
     }
-    
+
     return min;
   }
 
-  double calc_max(const std::vector<double>& toCalculateFrom)
+  double calc_max(const std::vector<double> &toCalculateFrom)
   {
     if (toCalculateFrom.size() == 0)
     {
@@ -73,41 +73,40 @@ namespace math_utility
 
     double max{toCalculateFrom.at(0)};
 
-    for (const auto& number : toCalculateFrom)
+    for (const auto &number : toCalculateFrom)
     {
       max = std::max(number, max);
     }
-    
+
     return max;
   }
 
-  std::vector<double> create_sorted_sequence(const std::vector<double>& toBeCopied)
+  std::vector<double> create_sorted_sequence(const std::vector<double> &toBeCopied)
   {
     auto toBeSorted = toBeCopied;
     std::sort(toBeSorted.begin(), toBeSorted.end());
     return toBeSorted;
   }
 
-  double get_median_from(const std::vector<double>& toExtractMedianFrom)
+  double get_median_from(const std::vector<double> &toExtractMedianFrom)
   {
     std::vector<double> sorted = create_sorted_sequence(toExtractMedianFrom);
-    
+
     int size = sorted.size();
     bool isSizeOdd = size % 2 == 0;
 
     int middleIndex = size / 2;
-    int middleValue = sorted.at(middleIndex);
+    double middleValue = sorted.at(middleIndex);
 
     if (!isSizeOdd)
     {
       return middleValue;
     }
-    else 
+    else
     {
-      double bottomMiddle = sorted.at(middleIndex - 1);
-      return (middleValue + bottomMiddle) / 2;
+      double bottomMiddle = sorted.at(middleIndex - 1);      
+      return (middleValue + bottomMiddle) / 2.0;
     }
   }
 
 }
-
